+++ src/util/test/mock_fetch.ts
@@ -1,4 +1,4 @@
-class RequestMock implements Partial<Request> {
+export class RequestMock implements Partial<Request> {
     public readonly cache: RequestCache;
     public readonly headers: Headers = new Headers();
     public readonly method?: string;
@@ -16,19 +16,24 @@ class RequestMock implements Partial<Request> {
     }
 }
 
-class AbortController {
+class AbortControllerMock {
     public signal: AbortSignal;
 
     public abort() {}
 }
 
-export function setupFetchMock() {
-    global.AbortController = AbortController;
-    global.Request = RequestMock as unknown as typeof Request;
-
-    global.fetch = jest.fn(() => {
-        return Promise.resolve(<Response>{
-            json: () => null,
-        });
+export function setupFetchMock(): jest.Mock<Promise<Response>, [input: RequestInfo | URL, init?: RequestInit], any> {
+    const fetchMock = jest.fn(async (_input: RequestInfo | URL, _init?: RequestInit): Promise<Response> => {
+        return <Response>{
+            json: async () => null,
+            arrayBuffer: async () => null,
+            text: async () => null,
+        };
     });
+
+    global.AbortController = AbortControllerMock;
+    global.Request = RequestMock as unknown as typeof Request;
+    global.fetch = fetchMock;
+
+    return fetchMock;
 }